SERVER - Refactor Payment, Product, Product Review, Role, Shipping and User

+++ server/src/main/java/com/shoppingsystem/shopping_system/user/service/UserServiceImpl.java
@@ -1,8 +1,11 @@
 package com.shoppingsystem.shopping_system.user.service;
 
+import com.shoppingsystem.shopping_system.auth.exceptions.InvalidCredentialsException;
 import com.shoppingsystem.shopping_system.role.model.Role;
 import com.shoppingsystem.shopping_system.role.service.RoleService;
+import com.shoppingsystem.shopping_system.user.dto.UserRequest;
 import com.shoppingsystem.shopping_system.user.dto.UserResponse;
+import com.shoppingsystem.shopping_system.user.exception.NoUserFoundException;
 import com.shoppingsystem.shopping_system.user.exception.UserNotFoundException;
 import com.shoppingsystem.shopping_system.user.model.User;
 import com.shoppingsystem.shopping_system.user.repository.UserRepository;
@@ -10,8 +13,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Service;
 
-import java.util.LinkedList;
+import java.time.Instant;
+import java.util.Date;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -56,6 +62,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Long userId) {
+        if(userId == null) {
+            throw new InvalidCredentialsException("User ID must not be null");
+        }
+
         userRepository.deleteById(userId);
     }
 
@@ -67,14 +77,39 @@ public class UserServiceImpl implements UserService {
     @Override
     public List<UserResponse> findAll() {
         List<User> userList = userRepository.findAll();
-        List<UserResponse> userResponses = new LinkedList<>();
-        for(User user : userList) {
-            Role role = roleService.findById(user.getRole().getId());
-            UserResponse userResponse = new UserResponse(user.getId(), user.getEmail(), "active", user.getFirstName(),
-                    user.getLastName(), user.getPasswordHash(), role.getId(), role.getName());
-            userResponses.add(userResponse);
+        if(userList.isEmpty()) {
+            throw new NoUserFoundException("No user found");
         }
-        return userResponses;
+
+        Map<Integer, Role> roleMap = generateRoleMap(userList);
+
+        return userList.stream()
+                .map(user -> {
+                    Role role = roleMap.get(user.getRole().getId());
+                    return new UserResponse(user.getId(), user.getEmail(), "active", user.getFirstName(),
+                            user.getLastName(), user.getPasswordHash(), role.getId(), role.getName());
+                }).toList();
+    }
+
+    private Map<Integer, Role> generateRoleMap(List<User> userList) {
+        List<Long> userIds = userList.stream().map(User::getId).toList();
+        List<Role> roleList = roleService.findRolesByUserIds(userIds);
+        return roleList.stream().collect(Collectors.toMap(
+                Role::getId, role -> role
+                ));
+    }
+
+    public void updateUser(UserRequest userRequest) {
+        User existingUser = userRepository.findById(userRequest.getUserId())
+                .orElseThrow(() -> new UserNotFoundException("Couldn't find user with ID - " + userRequest.getUserId()));
+        existingUser.setEmail(userRequest.getEmail());
+        existingUser.setRole(roleService.findById(userRequest.getRoleId()));
+        existingUser.setFirstName(userRequest.getFirstName());
+        existingUser.setLastName(userRequest.getLastName());
+        existingUser.setUpdatedAt(Date.from(Instant.now()));
+        updatePassword(userRequest.getUserId(), userRequest.getPassword());
+
+        userRepository.save(existingUser);
     }
 
     public void updatePassword(Long userId, String newPassword) {
